moving functions to libnative
