+++ src/org/GreenTeaScript/LibGreenTea.java
@@ -87,9 +87,9 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		else if(Func.Is(NativeMethodFunc)) {
 			/*local*/Object[] MethodArguments = new Object[Params.length-1];
 			LibGreenTea.ArrayCopy(Params, 1, MethodArguments, 0, MethodArguments.length);
-			return LibGreenTea.ApplyMethod(Func, Params[0], MethodArguments);
+			return LibNative.ApplyMethod(Func, Params[0], MethodArguments);
 		}
-		return LibGreenTea.ApplyMethod(Func, null, Params);
+		return LibNative.ApplyMethod(Func, null, Params);
 	}
 
 	public static Object InvokeOverridedMethod(long FileLine, GtNameSpace NameSpace, GtFunc Func, Object[] Arguments) {
@@ -394,8 +394,6 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		System.arraycopy(src, srcPos, dest, destPos, length);
 	}
 
-	
-
 	public final static GtFunc SetNativeMethod(GtFunc NativeFunc, Method JavaMethod) {
 		/*local*/int FuncFlag = GreenTeaUtils.NativeFunc;
 		if(!Modifier.isStatic(JavaMethod.getModifiers())) {
@@ -420,55 +418,8 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		return SetNativeMethod(new GtFunc(0, JavaMethod.getName(), 0, TypeList), JavaMethod);
 	}
 
-	public final static Class<?> LoadNativeClass(String ClassName) throws ClassNotFoundException {
-		try {
-			return Class.forName("org.GreenTeaScript." + ClassName);
-		}
-		catch(ClassNotFoundException e) {
-		}
-		return Class.forName(ClassName);
-	}
-	
-	public final static Method LoadNativeMethod(GtType ContextType, String FullName, boolean StaticMethodOnly) {
-		/*local*/Method FoundMethod = null;
-		int Index = FullName.lastIndexOf(".");
-		if(Index == -1) {
-			return null;
-		}
-		try {
-			/*local*/String FuncName = FullName.substring(Index+1);
-			/*local*/Class<?> NativeClass = LoadNativeClass(FullName.substring(0, Index));
-			Method[] Methods = NativeClass.getDeclaredMethods();
-			assert(Methods != null);
-			for(int i = 0; i < Methods.length; i++) {
-				if(LibGreenTea.EqualsString(FuncName, Methods[i].getName())) {
-					if(!Modifier.isPublic(Methods[i].getModifiers())) {
-						continue;
-					}
-					if(StaticMethodOnly && !Modifier.isStatic(Methods[i].getModifiers())) {
-						continue;
-					}
-					if(ContextType.IsFuncType() && !LibNative.MatchNativeMethod(ContextType.TypeParams, Methods[i])) {
-						continue;
-					}
-					if(FoundMethod != null) {
-						LibGreenTea.VerboseLog(GreenTeaUtils.VerboseUndefined, "overloaded method: " + FullName);
-						return FoundMethod; // return the first one
-					}
-					FoundMethod = Methods[i];
-				}
-			}
-			if(FoundMethod == null) {
-				LibGreenTea.VerboseLog(GreenTeaUtils.VerboseUndefined, "undefined method: " + FullName + " for " + ContextType);
-			}
-		} catch (ClassNotFoundException e) {
-				LibGreenTea.VerboseLog(GreenTeaUtils.VerboseException, e.toString());			
-		}
-		return FoundMethod;
-	}
-
 	public final static boolean ImportNativeMethod(GtNameSpace NameSpace, GtFunc NativeFunc, String FullName) {
-		Method JavaMethod = LibGreenTea.LoadNativeMethod(NativeFunc.GetFuncType(), FullName, false);
+		Method JavaMethod = LibNative.LoadNativeMethod(NativeFunc.GetFuncType(), FullName, false);
 		if(JavaMethod != null) {
 			LibGreenTea.SetNativeMethod(NativeFunc, JavaMethod);
 			if(NativeFunc.GetReturnType().IsVarType()) {
@@ -487,113 +438,16 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		return false;
 	}
 	
-	public final static Object ImportNativeObject(GtNameSpace NameSpace, String PackageName) {
-		LibGreenTea.VerboseLog(VerboseNative, "importing " + PackageName);
-//ifdef JAVA
-		try {
-			/*local*/Class<?> NativeClass = LibGreenTea.LoadNativeClass(PackageName);
-			try {
-				Method LoaderMethod = NativeClass.getMethod("ImportGrammar", GtNameSpace.class, Class.class);
-				LoaderMethod.invoke(null, NameSpace, NativeClass);
-			} catch (Exception e) {  // naming
-			}
-			return LibNative.GetNativeType(NativeClass);
-		} catch (ClassNotFoundException e) {
-		}
-		int Index = PackageName.lastIndexOf(".");
-		if(Index == -1) {
-			return null;
-		}
-		try {
-			/*local*/Class<?> NativeClass = LoadNativeClass(PackageName.substring(0, Index));
-			return ImportStaticObject(NameSpace.Context, NativeClass, PackageName.substring(Index+1));
-		} catch (ClassNotFoundException e) {
-		}
-//endif VAJA
-		return null;
-	}
-
 	public final static void LoadNativeConstructors(GtParserContext Context, GtType ClassType, ArrayList<GtFunc> FuncList) {
-		/*local*/boolean TransformedResult = false;
-		Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
-//		GtParserContext Context = ClassType.Context;
-		Constructor<?>[] Constructors = NativeClass.getDeclaredConstructors();
-		if(Constructors != null) {
-			for(int i = 0; i < Constructors.length; i++) {
-				if(!Modifier.isPublic(Constructors[i].getModifiers())) {
-					continue;
-				}
-				/*local*/ArrayList<GtType> TypeList = new ArrayList<GtType>();
-				TypeList.add(ClassType);
-				/*local*/Class<?>[] ParamTypes = Constructors[i].getParameterTypes();
-				if(ParamTypes != null) {
-					for(int j = 0; j < ParamTypes.length; j++) {
-						TypeList.add(LibNative.GetNativeType(ParamTypes[j]));
-					}
-				}
-				GtFunc Func = new GtFunc(ConstructorFunc, ClassType.ShortName, 0, TypeList);
-				Func.SetNativeMethod(0, Constructors[i]);
-				Context.RootNameSpace.AppendConstructor(ClassType, Func, null);
-				FuncList.add(Func);
-				TransformedResult = true;
-			}
-		}
-		if(!TransformedResult) {
-			Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.ConstructorSymbol()), null);
-		}
+		LibNative.LoadNativeConstructors(Context, ClassType, FuncList);
 	}
-
+	
 	public final static GtFunc LoadNativeField(GtParserContext Context, GtType ClassType, String FieldName, boolean GetSetter) {
-		try {
-			Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
-			Field NativeField = NativeClass.getField(FieldName);
-			if(Modifier.isPublic(NativeField.getModifiers())) {
-				ArrayList<GtType> TypeList = new ArrayList<GtType>();
-				TypeList.add(LibNative.GetNativeType(NativeField.getType()));
-				TypeList.add(ClassType);
-				GtFunc GetterNativeFunc = new GtFunc(GetterFunc, FieldName, 0, TypeList);
-				GetterNativeFunc.SetNativeMethod(0, NativeField);
-				Context.RootNameSpace.SetGetterFunc(ClassType, FieldName, GetterNativeFunc, null);
-				TypeList.clear();
-				TypeList.add(GtStaticTable.VoidType);
-				TypeList.add(ClassType);
-				TypeList.add(LibNative.GetNativeType(NativeField.getType()));
-				GtFunc SetterNativeFunc = new GtFunc(SetterFunc, FieldName, 0, TypeList);
-				SetterNativeFunc.SetNativeMethod(0, NativeField);
-				Context.RootNameSpace.SetSetterFunc(ClassType, FieldName, SetterNativeFunc, null);
-				return GetSetter ? SetterNativeFunc : GetterNativeFunc;
-			}
-		} catch (SecurityException e) {
-			LibGreenTea.VerboseException(e);
-		} catch (NoSuchFieldException e) {
-		}
-		Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.GetterSymbol(FieldName)), null);
-		Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.SetterSymbol(FieldName)), null); // for setter
-		return null;
+		return LibNative.LoadNativeField(Context, ClassType, FieldName, GetSetter);
 	}
 
-	public static Object NativeFieldValue(Object ObjectValue, Field NativeField) {
-		try {
-			Class<?> NativeType = NativeField.getType();
-			if(NativeType == long.class || NativeType == int.class || NativeType == short.class || NativeType == byte.class) {
-				return NativeField.getLong(ObjectValue);
-			}
-			if(NativeType == double.class || NativeType == float.class) {
-				return NativeField.getDouble(ObjectValue);
-			}
-			if(NativeType == boolean.class) {
-				return NativeField.getBoolean(ObjectValue);
-			}
-			if(NativeType == char.class) {
-				return String.valueOf(NativeField.getChar(ObjectValue));
-			}
-			return NativeField.get(ObjectValue);
-		} catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		} catch (SecurityException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
+	public final static void LoadNativeMethods(GtParserContext Context, GtType ClassType, String FuncName, ArrayList<GtFunc> FuncList) {
+		LibNative.LoadNativeMethods(Context, ClassType, FuncName, FuncList);
 	}
 
 	public static Object NativeFieldGetter(Object ObjectValue, Field NativeField) {
@@ -619,56 +473,6 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		return Value;
 	}
 
-	public static Object ImportStaticObject(GtParserContext Context, Class<?> NativeClass, String Symbol) {
-		try {
-			Field NativeField = NativeClass.getField(Symbol);
-			if(Modifier.isStatic(NativeField.getModifiers())) {
-				return NativeFieldValue(null, NativeField);
-			}
-		} catch (NoSuchFieldException e) {
-//			LibGreenTea.VerboseException(e);
-		}
-		GtPolyFunc PolyFunc = new GtPolyFunc(null);
-		Method[] Methods = NativeClass.getMethods();
-		for(int i = 0; i < Methods.length; i++) {
-			if(Methods[i].getName().equals(Symbol) && Modifier.isStatic(Methods[i].getModifiers())) {
-				PolyFunc.Append(Context, LibGreenTea.ConvertNativeMethodToFunc(Context, Methods[i]), null);
-			}
-		}
-		if(PolyFunc.FuncList.size() == 1) {
-			return PolyFunc.FuncList.get(0);
-		}
-		else if(PolyFunc.FuncList.size() != 0) {
-			return PolyFunc;
-		}
-		return null;
-	}
-
-	public static Object LoadNativeStaticFieldValue(GtParserContext Context, GtType ClassType, String Symbol) {
-		return ImportStaticObject(Context, (Class<?>)ClassType.TypeBody, Symbol);
-	}
-	
-	public final static void LoadNativeMethods(GtParserContext Context, GtType ClassType, String FuncName, ArrayList<GtFunc> FuncList) {
-		Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
-		Method[] Methods = NativeClass.getDeclaredMethods();
-		/*local*/boolean FoundMethod = false;
-		if(Methods != null) {
-			for(int i = 0; i < Methods.length; i++) {
-				if(LibGreenTea.EqualsString(FuncName, Methods[i].getName())) {
-					if(!Modifier.isPublic(Methods[i].getModifiers())) {
-						continue;
-					}
-					GtFunc NativeFunc = LibGreenTea.ConvertNativeMethodToFunc(Context, Methods[i]);
-					Context.RootNameSpace.AppendMethod(NativeFunc, null);
-					FuncList.add(NativeFunc);
-					FoundMethod = true;
-				}
-			}
-		}
-		if(!FoundMethod) {
-			Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, FuncName), null);
-		}
-	}
 
 	public final static Method LookupNativeMethod(Object Callee, String FuncName) {
 		if(FuncName != null) {
@@ -684,99 +488,8 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 		return null;
 	}
 
-	public final static Object ApplyMethod(GtFunc Func, Object Self, Object[] Params) {
-		try {
-//			System.err.println("** debug: " + Func.FuncBody);
-//			System.err.println("** debug: " + Self + ", Params.length=" + Params.length);
-			return ((Method)Func.FuncBody).invoke(Self, Params);
-		}
-		catch (InvocationTargetException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalArgumentException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
-	}
 
-	public final static Object ApplyMethod1(GtFunc Func, Object Self, Object Param) {
-		try {
-			return ((Method)Func.FuncBody).invoke(Self, Param);
-		}
-		catch (InvocationTargetException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalArgumentException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
-	}
 
-	public final static Object ApplyMethod2(GtFunc Func, Object Self, Object Param1, Object Param2) {
-		try {
-			return ((Method)Func.FuncBody).invoke(Self, Param1, Param2);
-		}
-		catch (InvocationTargetException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalArgumentException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
-	}
-
-	public final static Object ApplyMethod3(GtFunc Func, Object Self, Object Param1, Object Param2, Object Param3) {
-		try {
-			return ((Method)Func.FuncBody).invoke(Self, Param1, Param2, Param3);
-		}
-		catch (InvocationTargetException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalArgumentException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
-	}
-
-	public final static Object ApplyMethod4(GtFunc Func, Object Self, Object Param1, Object Param2, Object Param3, Object Param4) {
-		try {
-			return ((Method)Func.FuncBody).invoke(Self, Param1, Param2, Param3, Param4);
-		}
-		catch (InvocationTargetException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalArgumentException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		catch (IllegalAccessException e) {
-			LibGreenTea.VerboseException(e);
-		}
-		return null;
-	}
-
-	public final static long ApplyTokenFunc(GtFunc TokenFunc, Object TokenContext, String Text, long pos) {
-		return (Long)LibGreenTea.ApplyMethod3(TokenFunc, null, TokenContext, Text, pos);
-	}
-
-	public final static GtSyntaxTree ApplyParseFunc(GtFunc ParseFunc, GtNameSpace NameSpace, GtTokenContext TokenContext, GtSyntaxTree LeftTree, GtSyntaxPattern Pattern) {
-		return (GtSyntaxTree)LibGreenTea.ApplyMethod4(ParseFunc, null, NameSpace, TokenContext, LeftTree, Pattern);
-	}
-
-	public final static GtNode ApplyTypeFunc(GtFunc TypeFunc, GtTypeEnv Gamma, GtSyntaxTree ParsedTree, GtType ContextType) {
-		return (GtNode)LibGreenTea.ApplyMethod3(TypeFunc, null, Gamma, ParsedTree, ContextType);
-	}
 
 	public final static int ListSize(ArrayList<?> List) {
 		if(List == null) {
@@ -1135,7 +848,10 @@ public abstract class LibGreenTea implements GreenTeaConsts {
 			}
 			GtFunc Func = GtStaticTable.GetConverterFunc(ValueType, CastType, true);
 			if(Func != null) {
-				return LibGreenTea.ApplyMethod2(Func, null, CastType, Value);
+				Object[] Argvs = new Object[2];
+				Argvs[0] = CastType;
+				Argvs[1] = Value;
+				return LibNative.ApplyMethod(Func, null, Argvs);
 			}
 		}
 		return null;
