+++ src/org/GreenTeaScript/LibNative.java
@@ -38,20 +38,34 @@ package org.GreenTeaScript;
 //import java.lang.reflect.Constructor;
 //import java.lang.reflect.Field;
 //import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 //import java.util.ArrayList;
 //import java.util.Iterator;
 //import org.GreenTeaScript.Konoha.ArrayApi;
+import java.util.ArrayList;
 
 public class LibNative {
 	final static void DebugP(String s) {
 		//System.err.println("LibNative.DebugP: " + s);
 	}
+
 	// type
+	private final static Class<?> LoadNativeClass(String ClassName) throws ClassNotFoundException {
+		try {
+			return Class.forName("org.GreenTeaScript." + ClassName);
+		}
+		catch(ClassNotFoundException e) {
+		}
+		return Class.forName(ClassName);
+	}
 	public final static Class<?> GetClassOfValue(Object Value) {
 		return Value.getClass();
 	}
+
 	public final static GtType GetNativeType(Class<?> NativeClass) {
 		GtType NativeType = null;
 		NativeType = (/*cast*/GtType) GtStaticTable.ClassNameMap.GetOrNull(NativeClass.getCanonicalName());
@@ -85,7 +99,7 @@ public class LibNative {
 		return true;
 	}
 
-	public final static boolean MatchNativeMethod(GtType[] GreenTypeParams, Method JavaMethod) {
+	private final static boolean MatchNativeMethod(GtType[] GreenTypeParams, Method JavaMethod) {
 		if(!AcceptJavaType(GreenTypeParams[0], JavaMethod.getReturnType())) {
 			DebugP("return mismatched: " + GreenTypeParams[0] + ", " + JavaMethod.getReturnType() + " of " + JavaMethod);
 			return false;
@@ -119,4 +133,312 @@ public class LibNative {
 		return (ParamSize == 0);
 	}
 
+	public final static Method LoadNativeMethod(GtType ContextType, String FullName, boolean StaticMethodOnly) {
+		/*local*/Method FoundMethod = null;
+		int Index = FullName.lastIndexOf(".");
+		if(Index == -1) {
+			return null;
+		}
+		try {
+			/*local*/String FuncName = FullName.substring(Index+1);
+			/*local*/Class<?> NativeClass = LibNative.LoadNativeClass(FullName.substring(0, Index));
+			Method[] Methods = NativeClass.getDeclaredMethods();
+			assert(Methods != null);
+			for(int i = 0; i < Methods.length; i++) {
+				if(LibGreenTea.EqualsString(FuncName, Methods[i].getName())) {
+					if(!Modifier.isPublic(Methods[i].getModifiers())) {
+						continue;
+					}
+					if(StaticMethodOnly && !Modifier.isStatic(Methods[i].getModifiers())) {
+						continue;
+					}
+					if(ContextType.IsFuncType() && !LibNative.MatchNativeMethod(ContextType.TypeParams, Methods[i])) {
+						continue;
+					}
+					if(FoundMethod != null) {
+						LibGreenTea.VerboseLog(GreenTeaUtils.VerboseUndefined, "overloaded method: " + FullName);
+						return FoundMethod; // return the first one
+					}
+					FoundMethod = Methods[i];
+				}
+			}
+			if(FoundMethod == null) {
+				LibGreenTea.VerboseLog(GreenTeaUtils.VerboseUndefined, "undefined method: " + FullName + " for " + ContextType);
+			}
+		} catch (ClassNotFoundException e) {
+				LibGreenTea.VerboseLog(GreenTeaUtils.VerboseException, e.toString());			
+		}
+		return FoundMethod;
+	}
+
+	public static Object GetNativeFieldValue(Object ObjectValue, Field NativeField) {
+		try {
+			Class<?> NativeType = NativeField.getType();
+			if(NativeType == long.class || NativeType == int.class || NativeType == short.class || NativeType == byte.class) {
+				return NativeField.getLong(ObjectValue);
+			}
+			if(NativeType == double.class || NativeType == float.class) {
+				return NativeField.getDouble(ObjectValue);
+			}
+			if(NativeType == boolean.class) {
+				return NativeField.getBoolean(ObjectValue);
+			}
+			if(NativeType == char.class) {
+				return String.valueOf(NativeField.getChar(ObjectValue));
+			}
+			return NativeField.get(ObjectValue);
+		} catch (IllegalAccessException e) {
+			LibGreenTea.VerboseException(e);
+		} catch (SecurityException e) {
+			LibGreenTea.VerboseException(e);
+		}
+		return null;
+	}
+
+	public static Object ImportStaticObject(GtParserContext Context, Class<?> NativeClass, String Symbol) {
+		try {
+			Field NativeField = NativeClass.getField(Symbol);
+			if(Modifier.isStatic(NativeField.getModifiers())) {
+				return LibNative.GetNativeFieldValue(null, NativeField);
+			}
+		} catch (NoSuchFieldException e) {
+//			LibGreenTea.VerboseException(e);
+		}
+		GtPolyFunc PolyFunc = new GtPolyFunc(null);
+		Method[] Methods = NativeClass.getMethods();
+		for(int i = 0; i < Methods.length; i++) {
+			if(Methods[i].getName().equals(Symbol) && Modifier.isStatic(Methods[i].getModifiers())) {
+				PolyFunc.Append(Context, LibGreenTea.ConvertNativeMethodToFunc(Context, Methods[i]), null);
+			}
+		}
+		if(PolyFunc.FuncList.size() == 1) {
+			return PolyFunc.FuncList.get(0);
+		}
+		else if(PolyFunc.FuncList.size() != 0) {
+			return PolyFunc;
+		}
+		return null;
+	}
+
+	public static Object LoadNativeStaticFieldValue(GtParserContext Context, GtType ClassType, String Symbol) {
+		return LibNative.ImportStaticObject(Context, (Class<?>)ClassType.TypeBody, Symbol);
+	}
+	
+	public final static Object ImportNativeObject(GtNameSpace NameSpace, String PackageName) {
+		LibGreenTea.VerboseLog(GreenTeaUtils.VerboseNative, "importing " + PackageName);
+		try {
+			/*local*/Class<?> NativeClass = LibNative.LoadNativeClass(PackageName);
+			try {
+				Method LoaderMethod = NativeClass.getMethod("ImportGrammar", GtNameSpace.class, Class.class);
+				LoaderMethod.invoke(null, NameSpace, NativeClass);
+			} catch (Exception e) {  // naming
+			}
+			return LibNative.GetNativeType(NativeClass);
+		} catch (ClassNotFoundException e) {
+		}
+		int Index = PackageName.lastIndexOf(".");
+		if(Index == -1) {
+			return null;
+		}
+		try {
+			/*local*/Class<?> NativeClass = LibNative.LoadNativeClass(PackageName.substring(0, Index));
+			return ImportStaticObject(NameSpace.Context, NativeClass, PackageName.substring(Index+1));
+		} catch (ClassNotFoundException e) {
+		}
+		return null;
+	}
+
+	public final static void LoadNativeConstructors(GtParserContext Context, GtType ClassType, ArrayList<GtFunc> FuncList) {
+		/*local*/boolean TransformedResult = false;
+		Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
+//		GtParserContext Context = ClassType.Context;
+		Constructor<?>[] Constructors = NativeClass.getDeclaredConstructors();
+		if(Constructors != null) {
+			for(int i = 0; i < Constructors.length; i++) {
+				if(!Modifier.isPublic(Constructors[i].getModifiers())) {
+					continue;
+				}
+				/*local*/ArrayList<GtType> TypeList = new ArrayList<GtType>();
+				TypeList.add(ClassType);
+				/*local*/Class<?>[] ParamTypes = Constructors[i].getParameterTypes();
+				if(ParamTypes != null) {
+					for(int j = 0; j < ParamTypes.length; j++) {
+						TypeList.add(LibNative.GetNativeType(ParamTypes[j]));
+					}
+				}
+				GtFunc Func = new GtFunc(GreenTeaConsts.ConstructorFunc, ClassType.ShortName, 0, TypeList);
+				Func.SetNativeMethod(0, Constructors[i]);
+				Context.RootNameSpace.AppendConstructor(ClassType, Func, null);
+				FuncList.add(Func);
+				TransformedResult = true;
+			}
+		}
+		if(!TransformedResult) {
+			Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.ConstructorSymbol()), null);
+		}
+	}
+	
+	public final static GtFunc LoadNativeField(GtParserContext Context, GtType ClassType, String FieldName, boolean GetSetter) {
+		try {
+			Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
+			Field NativeField = NativeClass.getField(FieldName);
+			if(Modifier.isPublic(NativeField.getModifiers())) {
+				ArrayList<GtType> TypeList = new ArrayList<GtType>();
+				TypeList.add(LibNative.GetNativeType(NativeField.getType()));
+				TypeList.add(ClassType);
+				GtFunc GetterNativeFunc = new GtFunc(GreenTeaConsts.GetterFunc, FieldName, 0, TypeList);
+				GetterNativeFunc.SetNativeMethod(0, NativeField);
+				Context.RootNameSpace.SetGetterFunc(ClassType, FieldName, GetterNativeFunc, null);
+				TypeList.clear();
+				TypeList.add(GtStaticTable.VoidType);
+				TypeList.add(ClassType);
+				TypeList.add(LibNative.GetNativeType(NativeField.getType()));
+				GtFunc SetterNativeFunc = new GtFunc(GreenTeaConsts.SetterFunc, FieldName, 0, TypeList);
+				SetterNativeFunc.SetNativeMethod(0, NativeField);
+				Context.RootNameSpace.SetSetterFunc(ClassType, FieldName, SetterNativeFunc, null);
+				return GetSetter ? SetterNativeFunc : GetterNativeFunc;
+			}
+		} catch (SecurityException e) {
+			LibGreenTea.VerboseException(e);
+		} catch (NoSuchFieldException e) {
+		}
+		Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.GetterSymbol(FieldName)), null);
+		Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, GreenTeaUtils.SetterSymbol(FieldName)), null); // for setter
+		return null;
+	}
+
+	public final static void LoadNativeMethods(GtParserContext Context, GtType ClassType, String FuncName, ArrayList<GtFunc> FuncList) {
+		Class<?> NativeClass = (Class<?>)ClassType.TypeBody;
+		Method[] Methods = NativeClass.getDeclaredMethods();
+		/*local*/boolean FoundMethod = false;
+		if(Methods != null) {
+			for(int i = 0; i < Methods.length; i++) {
+				if(LibGreenTea.EqualsString(FuncName, Methods[i].getName())) {
+					if(!Modifier.isPublic(Methods[i].getModifiers())) {
+						continue;
+					}
+					GtFunc NativeFunc = LibGreenTea.ConvertNativeMethodToFunc(Context, Methods[i]);
+					Context.RootNameSpace.AppendMethod(NativeFunc, null);
+					FuncList.add(NativeFunc);
+					FoundMethod = true;
+				}
+			}
+		}
+		if(!FoundMethod) {
+			Context.RootNameSpace.SetUndefinedSymbol(GreenTeaUtils.ClassSymbol(ClassType, FuncName), null);
+		}
+	}
+
+	// Method
+	public final static Object ApplyMethod(GtFunc Func, Object Self, Object[] Params) {
+		try {
+//			System.err.println("** debug: " + Func.FuncBody);
+//			System.err.println("** debug: " + Self + ", Params.length=" + Params.length);
+			return ((Method)Func.FuncBody).invoke(Self, Params);
+		}
+		catch (InvocationTargetException e) {
+			LibGreenTea.VerboseException(e);
+		}
+		catch (IllegalArgumentException e) {
+			LibGreenTea.VerboseException(e);
+		}
+		catch (IllegalAccessException e) {
+			LibGreenTea.VerboseException(e);
+		}
+		return null;
+	}
+
+	public final static long ApplyTokenFunc(GtFunc TokenFunc, Object TokenContext, String Text, long pos) {
+		Object[] Argvs = new Object[3];
+		Argvs[0] = TokenContext;
+		Argvs[1] = Text;
+		Argvs[2] = pos;
+		return (Long)LibNative.ApplyMethod(TokenFunc, null, Argvs);
+	}
+
+	public final static GtSyntaxTree ApplyParseFunc(GtFunc ParseFunc, GtNameSpace NameSpace, GtTokenContext TokenContext, GtSyntaxTree LeftTree, GtSyntaxPattern Pattern) {
+		Object[] Argvs = new Object[4];
+		Argvs[0] = NameSpace;
+		Argvs[1] = TokenContext;
+		Argvs[2] = LeftTree;
+		Argvs[3] = Pattern;
+		return (GtSyntaxTree)LibNative.ApplyMethod(ParseFunc, null, Argvs);
+	}
+
+	public final static GtNode ApplyTypeFunc(GtFunc TypeFunc, GtTypeEnv Gamma, GtSyntaxTree ParsedTree, GtType ContextType) {
+		Object[] Argvs = new Object[3];
+		Argvs[0] = Gamma;
+		Argvs[1] = ParsedTree;
+		Argvs[2] = ContextType;
+		return (GtNode)LibNative.ApplyMethod(TypeFunc, null, Argvs);
+	}
+
+//	public final static Object ApplyMethod1(Method FuncBody, Object Self, Object Param) {
+//		try {
+//			return FuncBody.invoke(Self, Param);
+//		}
+//		catch (InvocationTargetException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalArgumentException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalAccessException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		return null;
+//	}
+
+
+//	public final static Object ApplyMethod2(Method FuncBody, Object Self, Object Param1, Object Param2) {
+//		try {
+//			return FuncBody.invoke(Self, Param1, Param2);
+//		}
+//		catch (InvocationTargetException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalArgumentException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalAccessException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		return null;
+//	}
+//
+//	public final static Object ApplyMethod3(Method FuncBody, Object Self, Object Param1, Object Param2, Object Param3) {
+//		try {
+//			return FuncBody.invoke(Self, Param1, Param2, Param3);
+//		}
+//		catch (InvocationTargetException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalArgumentException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalAccessException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		return null;
+//	}
+//
+//	public final static Object ApplyMethod4(Method FuncBody, Object Self, Object Param1, Object Param2, Object Param3, Object Param4) {
+//		try {
+//			return FuncBody.invoke(Self, Param1, Param2, Param3, Param4);
+//		}
+//		catch (InvocationTargetException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalArgumentException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		catch (IllegalAccessException e) {
+//			LibGreenTea.VerboseException(e);
+//		}
+//		return null;
+//	}
+
+	
+	
+	
 }
