+++ test/bugs/test5.green
@@ -1,0 +1,33 @@
+let ArrayList = import java.util.ArrayList;
+let fibodata = new ArrayList();
+
+int fibo(int n) {
+        if(n < fibodata.size()) {
+                return fibodata.get(n - 1);
+        }else if(n < fibodata.size() - 1) {
+                fibo(n - 1);
+        }
+
+        fibodata.add(fibo(n - 1) + fibo(n - 2));
+        return fibodata.get(n - 1);
+}
+
+int fibo_init() {
+        fibodata.add(1);
+        fibodata.add(1);
+
+        return 0;
+}
+
+@Export int main() {
+        int maxnum = 100;
+        fibo_init();
+
+        int i = 1;
+        while(i <= maxnum) {
+                println("f(" + i + ") = " + fibo(i));
+                i = i + 1;
+        }
+        return 0;
+}
+
